# backend/internal/handler/openapi.yaml
openapi: 3.0.3
info:
  title: 影策 HTTP API
  version: "1.0.0"
  description: |
    路径前缀是 `/api`。前后端一起发布，不按路径做 API 版本。

    业务 JSON 信封为 `{ code, data, msg, reason }`。`code === 0` 表示成功；失败时 HTTP status 与
    `code` 都表达失败。`code` 默认等于 HTTP 状态；同一状态下的细分原因使用 `status*100+n`
    （例如 40301 配额、42901 限流）。机器判断请用 `reason`，不要解析 `msg`。

    TypeScript 调用面仍是 `web/src/services/api` 的 `http` 模块，不从本规范生成客户端。
servers:
  - url: /api
security:
  - sessionCookie: []
...
paths:
...
  /admin/resources/{id}/file:
    get:
      tags: [admin-storage]
      summary: 管理员解析 CDN 地址、预览或下载资源
      description: 校验管理员权限后按资源所属用户的存储配置交付。云端资源只使用 CDN，缺少 CDN 或不可用时返回 503；仅本地资源返回文件流。resolve=1 返回包含 data.url 的 JSON 信封，本地资源的 url 为空。
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: query
          name: direct
          schema: { type: string, enum: ["1"] }
          description: 不绕过 CDN 交付策略，不签发源站回退地址。
        - in: query
          name: proxy
          schema: { type: string, enum: ["1"] }
          description: 不再启用云端媒体代理，仍使用 CDN 交付策略。
        - in: query
          name: resolve
          schema: { type: string, enum: ["1"] }
          description: 只解析读取地址，不传输媒体；签名地址禁止持久化。
        - in: query
          name: download
          schema: { type: string, enum: ["1"] }
        - in: query
          name: filename
          schema: { type: string }
          description: 下载文件名；七牛 CDN 通过签名 attname 设置，本地文件通过 Content-Disposition 设置。
        - in: header
          name: Range
          schema: { type: string }
      responses:
        "200":
          description: 本地文件流，或 resolve=1 时的 JSON 信封（data.url 为 CDN 地址或空字符串）。
        "206":
          description: 部分文件流
        "307":
          description: CDN 跳转；签名跳转禁止缓存。
          headers:
            Location:
              schema: { type: string, format: uri }
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
...
components:
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: open_ai_canvas_session
  schemas:
...
    Envelope:
      type: object
      required: [code, msg]
      properties:
        code:
          type: integer
          description: 0 成功；失败时默认等于 HTTP 状态，或 status*100+n 的细分码
        data: {}
        msg:
          type: string
        reason:
          type: string
          description: 仅失败时出现。稳定机器可读原因。
    ErrorEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            code:
              type: integer
              enum: [400, 401, 403, 404, 409, 429, 500, 502, 503, 504, 40301, 40901, 42901]
            data:
              nullable: true
            reason:
              type: string
              enum:
                - invalid_argument
                - unauthorized
                - forbidden
                - not_found
                - conflict
                - failed_precondition
                - quota_exceeded
                - rate_limited
                - unavailable
                - timeout
                - internal
                - bad_gateway
                - upstream_dns_failed
                - model_capability_not_supported
                - model_price_not_configured
                - model_route_unavailable
                - provider_request_failed
                - model_catalog_mismatch
                - invalid_model_selection
...
  responses:
...
    Error:
      description: 业务失败信封
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"